+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IFileVersionEventCollection.cs
@@ -4,7 +4,7 @@
     /// Collection of file version events
     /// </summary>
     [ConcreteType(typeof(FileVersionEventCollection))]
-    public interface IFileVersionEventCollection : IDataModelCollection<IFileVersionEvent>, IDataModelCollectionLoad<IFileVersionEvent>
+    public interface IFileVersionEventCollection : IDataModelCollection<IFileVersionEvent>, IDataModelCollectionLoad<IFileVersionEvent>, ISupportModules<IFileVersionEventCollection>
     {
 
     }