+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/IFolderCollection.cs
@@ -9,7 +9,7 @@ namespace PnP.Core.Model.SharePoint
     /// Public interface to define a collection of Folder objects
     /// </summary>
     [ConcreteType(typeof(FolderCollection))]
-    public interface IFolderCollection : IQueryable<IFolder>, IAsyncEnumerable<IFolder>, IDataModelCollection<IFolder>, IDataModelCollectionLoad<IFolder>, IDataModelCollectionDeleteByGuidId
+    public interface IFolderCollection : IQueryable<IFolder>, IAsyncEnumerable<IFolder>, IDataModelCollection<IFolder>, IDataModelCollectionLoad<IFolder>, IDataModelCollectionDeleteByGuidId, ISupportModules<IFolderCollection>
     {
         #region Add